refac: modularização do código
