--- assets/websocket/payload.json
@@ -1,20 +1,0 @@
-[
-    {
-        "action":"newOrder",
-        "params":{
-            "table": "${localStorage.getItem('tableName')}",
-            "item": "${item.id}",
-            "valor": "${item.valor}",
-            "quantidade": "${item.quantidade}"
-        }
-    },
-    {
-        "action":"newOrder",
-        "params":{
-            "table": "${localStorage.getItem('tableName')}",
-            "item": "${item.id}",
-            "valor": "${item.valor}",
-            "quantidade": "${item.quantidade}"
-        }
-    }
-]
