+++ assets/websocket/server.js
@@ -5,9 +5,10 @@ const WebSocketServer = require('websocket').server;
 const http = require('http');
 require('dotenv').config();
 
-// Inicializando listas para armazenar clientes conectados e cozinha conectada
-const clientsConnected = [];
-const kitchenConnected = [];
+// Imports de módulos locais
+const { clientsConnected, kitchenConnected, getIndexByConnection } = require('./connections');
+const utils = require('./utils');
+const handlers = require('./handlers');
 
 // Cria o servidor HTTP para lidar com solicitações que não são do WebSocket
 const server = http.createServer((request, response) => {
@@ -40,12 +41,22 @@ wsServer.on('request', (request) => {
         try {
             // Tentando analisar a mensagem JSON
             const data = JSON.parse(message.utf8Data);
+            const validationError = utils.validateMessage(data);
+            if (validationError) {
+                // Se houver um erro de validação, o tratamos aqui
+                const errorMessage = utils.formatMessage("erro", validationError);
+                connection.sendUTF(errorMessage);
+                console.log(errorMessage);
+                return;
+            }
             const action = data.action;
+
+            // swtich cases para cada action
             switch (action) {
                 case "login":
                     // Tratar ação de login
-                    doLogin(data.params.table, connection);
-                    const answerMessage = formatMessage("loginAnswer", 'success');
+                    handlers.doLogin(data.params.table, connection);
+                    const answerMessage = utils.formatMessage("loginAnswer", 'success');
                     connection.sendUTF(answerMessage);
                     break;
                 case "newOrder":
@@ -53,51 +64,9 @@ wsServer.on('request', (request) => {
             }
         } catch (e) {
             // Tratamento de erros para mensagens malformadas
-            const errorMessage = formatMessage("erro", 'Formato da mensagem inválido...');
+            const errorMessage = utils.formatMessage("erro", 'Formato da mensagem inválido...');
             connection.sendUTF(errorMessage);
             console.log(errorMessage);
         }
     });
 });
-
-// Função para encontrar o índice do cliente na lista de clientes conectados
-const getIndexByConnection = (connection) => clientsConnected.findIndex((valor) => connection == valor);
-
-// Função para lidar com a ação de login
-const doLogin = (table, connection) => {
-    const index = getIndexByConnection(connection);
-    if (index === -1) {
-        // Erro caso a conexão não seja encontrada
-        const mensagem = formatMessage("erro", "Erro ao efetuar login");
-        connection.sendUTF(mensagem);
-        console.log(`Erro ao efetuar login, MESA ${table}`);
-    } else {
-        // Conexão de cozinha ou mesa
-        if (table === "kitchen") { 
-            kitchenConnected = connection;
-        } else {
-            console.log(`Mesa online ${table}`);
-        }
-    }
-};
-
-// Função para criar um pedido
-// 'tableName'
-// item.id
-// item.valor
-// item.quantidade
-// const createdOrder = (table, )
-
-// Função para formatar mensagens de resposta
-const formatMessage = (action, data) => {
-    let mensagem;
-    switch (action) {
-        case 'erro':
-        case 'loginAnswer':
-            // Formatação básica para mensagens de erro ou resposta de login
-            mensagem = { "action": action, "params": { "msg": data } };
-            break;
-        // Outros casos podem ser adicionados aqui
-    }
-    return JSON.stringify(mensagem); // Convertendo o objeto em string JSON
-};
